pyproject.toml
[tool.poetry]
name = "minstral autogen"
version = "0.1.0"
description = "description"
authors = ["richard"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
pyautogen = "^0.1.10"
openai = "^0.28.1"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
datasets = "^2.14.5"
ipython = "^8.16.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
